requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.9.0
ollama==0.5.1
python-dateutil==2.8.2
email-validator>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-mock>=3.11.0
